requirements.txt
asgiref==3.8.1
bleach==6.1.0
boto3==1.34.131
botocore==1.34.131
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
celery==5.4.0
Django==4.2
django-cors-headers==4.4.0
django-crontab==0.7.1
django-environ==0.11.2
django-filter==24.2
django-fsm==3.0.0
django-model-utils==4.5.1
django-phonenumber-field==7.3.0
django-render-block==0.9.2
django-storages==1.14.3
django-templated-email==3.0.1
django-timezone-field==6.1.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-yasg==1.21.7
drf-yasg2==1.19.4
drf_api_logger==1.1.15
fcm-django==2.1.0
firebase-admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
jwt==1.3.1
Levenshtein==0.26.0
MarkupSafe==2.1.5
msgpack==1.0.8
packaging==24.1
phonenumbers==8.13.39
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.0
pytz==2024.1
PyYAML==6.0.1
RapidFuzz==3.10.0
requests==2.32.3
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
unicef-djangolib==0.7
unicef-restlib==0.7
uritemplate==4.1.1
urllib3==2.2.2
webencodings==0.5.1
django-jazzmin
django-import-export
django-celery-results
setuptools
psycopg2-binary==2.9.9
